Bubbles class fully functioning + modified Enemy
The bubbles enemies are working as they should, and they work on both directions, also, i modified the Enemy class so that it has the movement functions as base, so that any new enemy that needs to do the same behaviours have a class to use. Also there is a minor change in Vmedusa script, refering to the Destroy after dissapearing completely of the screen.